Change model field types

+++ models/user.go
@@ -7,12 +7,12 @@ import (
 )
 
 type User struct {
-	Id        uint64 `json:"id" db:"user_id"`
+	Id        uint   `json:"id" db:"user_id"`
 	FirstName string `json:"first_name" db:"first_name"`
 	LastName  string `json:"last_name" db:"last_name"`
 	Email     string `json:"email" db:"email"`
 	Password  string `json:"password" db:"password"`
-	RoleId    uint64 `json:"role_id" db:"role_id"`
+	RoleId    uint   `json:"role_id" db:"role_id"`
 }
 
 func (usr User) MarshalJSON() ([]byte, error) {
@@ -20,7 +20,7 @@ func (usr User) MarshalJSON() ([]byte, error) {
 		FirstName string `json:"first_name"`
 		LastName  string `json:"last_name"`
 		Email     string `json:"email"`
-		RoleID    uint64 `json:"role_id"`
+		RoleID    uint   `json:"role_id"`
 	}
 
 	tmp.FirstName = usr.FirstName
